Factor out api-related methods to please Rubocop
The class finally tipped the scales and Rubocop didn't like that.
This is a quick fix until some more substantial refactoring takes place.
